fix: entity list structure
